safari-600.7-branch/LayoutTests/fast/repaint/animation-after-layer-scroll-expected.txt

Test that animated gif on layer that is scrolled to view starts animating.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS layerTreeAsText.indexOf('repaint rect') > -1 is true
PASS successfullyParsed is true

TEST COMPLETE
